apps/api/requirements.txt
fastapi
uvicorn
python-dotenv
supabase
crewai
crewai-tools
langchain
# langchain-groq
langchain-google-genai
langchain-openai
pydantic
openai
litellm
